refactor competition table into components
